refactor: harden config constraints across all services
Systematic constraint tightening for Pydantic config models:

- Add min_length=1 to string fields that must not be empty (metrics
  host/path, seeder file_path, api host, GeoConfig database paths)
- Add field_validator on RelayOverrideTimeouts enforcing >= 0.1s
- Align PublishingConfig.timeout to ge=0.1 (was gt=0)
- Change monitor interval fields from int to float for consistency
  with BaseServiceConfig.interval
- Add RetryConfig.validate_max_delay ensuring max_delay >= initial_delay
- Add model_validator on ApiSourceConfig enforcing connect_timeout
  does not exceed timeout
- Add field_validator on DvmConfig.relays validating WebSocket URLs
  via Relay model

All fixes are backwards-compatible: they restrict constraints without
changing defaults, so no existing YAML config breaks.

+++ src/bigbrotr/services/monitor/service.py
@@ -376,7 +376,7 @@ class Monitor(
         See Also:
             ``fetch_relays_to_monitor``: The SQL query executed.
         """
-        monitored_before = int(self.chunk_progress.started_at) - self._config.discovery.interval
+        monitored_before = int(self.chunk_progress.started_at - self._config.discovery.interval)
         return await fetch_relays_to_monitor(self._brotr, monitored_before, networks)
 
     async def _check_chunk(
@@ -761,7 +761,7 @@ class Monitor(
         *,
         enabled: bool,
         relays: list[Relay],
-        interval: int,
+        interval: float,
         state_key: str,
         builder: EventBuilder,
         event_name: str,
